# api-docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Estrategia Educacional Code Challenge",
        "description": "A REST API implemented using Go programming language and Postgres database to manage operations linked to Github starred repositories and related tags.",
        "contact": {
            "name": "Icaro Ribeiro",
            "email": "icaroribeiro@hotmail.com"
        }
    },
    "servers": [{
        "url": "http://X.X.X.X:8080"
    }],
...
    "paths": {
...
        "/repositories/{repositoryId}": {
            "get": {
                "tags": [
                    "repository"
                ],
                "summary": "API endpoint designed to get a specific repository by its id.",
                "operationId": "GetRepository",
                "parameters": [{
                        "name": "repositoryId",
                        "in": "path",
                        "description": "The id of the repository.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "required": [
                                        "id",
                                        "name",
                                        "description",
                                        "url",
                                        "language"
                                    ],
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "url": {
                                            "type": "string"
                                        },
                                        "language": {
                                            "type": "string"
                                        },
                                        "tags": {
                                            "type":"array",
                                            "items":{
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Got an internal application error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{userId}/githubStarredRepositories": {
            "get": {
                "tags": [
                    "repository"
                ],
                "summary": "API endpoint designed to get the list of all user Github starred repositories.",
                "operationId": "GetAllUserGithubStarredRepositories",
                "parameters": [{
                        "name": "userId",
                        "in": "path",
                        "description": "The id of the user.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "required": [
                                            "id",
                                            "name",
                                            "description",
                                            "url",
                                            "language"
                                        ],
                                        "properties": {
                                            "id": {
                                                "type": "string"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "url": {
                                                "type": "string"
                                            },
                                            "language": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Got an internal application error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/users/{userId}/repositories": {
            "get": {
                "tags": [
                    "repository"
                ],
                "summary": "API endpoint designed to get the list of all user repositories.",
                "operationId": "GetAllUserRepositories",
                "parameters": [{
                        "name": "userId",
                        "in": "path",
                        "description": "The id of the user.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "required": [
                                            "id",
                                            "name",
                                            "description",
                                            "url",
                                            "language"
                                        ],
                                        "properties": {
                                            "id": {
                                                "type": "string"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "url": {
                                                "type": "string"
                                            },
                                            "language": {
                                                "type": "string"
                                            },
                                            "tags": {
                                                "type":"array",
                                                "items":{
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Got an internal application error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Error": {
                "type": "object",
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "type": "string"
                    }
                }
            }
        }
    }
}